+++ src/main/java/org/example/staystylish/domain/traveloutfit/controller/TravelOutfitController.java
@@ -5,11 +5,11 @@ import lombok.RequiredArgsConstructor;
 import org.example.staystylish.common.dto.response.ApiResponse;
 import org.example.staystylish.common.dto.response.PageResponse;
 import org.example.staystylish.common.security.UserPrincipal;
+import org.example.staystylish.domain.traveloutfit.code.TravelOutfitSuccessCode;
 import org.example.staystylish.domain.traveloutfit.dto.request.TravelOutfitRequest;
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitDetailResponse;
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitResponse;
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitSummaryResponse;
-import org.example.staystylish.domain.traveloutfit.exception.TravelOutfitSuccessCode;
 import org.example.staystylish.domain.traveloutfit.service.TravelOutfitService;
 import org.example.staystylish.domain.user.entity.Gender;
 import org.springframework.data.domain.Page;
