+++ src/test/java/org/example/staystylish/domain/traveloutfit/service/TravelOutfitServiceTest.java
@@ -19,6 +19,7 @@ import org.example.staystylish.domain.globalweather.client.WeatherApiClient;
 import org.example.staystylish.domain.globalweather.client.WeatherApiClient.Daily;
 import org.example.staystylish.domain.traveloutfit.ai.TravelAiClient;
 import org.example.staystylish.domain.traveloutfit.ai.TravelAiPromptBuilder;
+import org.example.staystylish.domain.traveloutfit.code.TravelOutfitErrorCode;
 import org.example.staystylish.domain.traveloutfit.dto.request.TravelOutfitRequest;
 import org.example.staystylish.domain.traveloutfit.dto.response.AiTravelJsonResponse;
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitDetailResponse;
@@ -26,7 +27,6 @@ import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitResp
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitResponse.CulturalConstraints;
 import org.example.staystylish.domain.traveloutfit.dto.response.TravelOutfitSummaryResponse;
 import org.example.staystylish.domain.traveloutfit.entity.TravelOutfit;
-import org.example.staystylish.domain.traveloutfit.exception.TravelOutfitErrorCode;
 import org.example.staystylish.domain.traveloutfit.repository.TravelOutfitRepository;
 import org.example.staystylish.domain.user.entity.Gender;
 import org.junit.jupiter.api.DisplayName;
